Added createMockStore test util
